add shirk file limit

+++ Lib_XDownloader/src/main/java/com/infinite/downloader/lru/LruDiskUsage.java
@@ -16,6 +16,7 @@ import java.util.concurrent.Executors;
  * Description: Class description
  */
 public abstract class LruDiskUsage implements DiskUsage {
+    private static int MIN_NOT_ALLOWED_DELETE_COUNT = 4;
 
     private final ExecutorService workerThread = Executors.newSingleThreadExecutor();
 
@@ -26,8 +27,16 @@ public abstract class LruDiskUsage implements DiskUsage {
 
     private void touchInBackground(File file) throws IOException {
         Files.setLastModifiedNow(file);
-        List<File> files = Files.getLruListFiles(file.getParentFile());
-        trim(files);
+        File parentFile = file.getParentFile();
+        if (parentFile != null) {
+            List<File> files = Files.getLruListFiles(parentFile);
+            int fileCount = files != null ? files.size() : 0;
+            DLogger.d("touch in background,dir file count is:" + fileCount);
+            //if the dir count is less than 5,not do trim,because maybe the latest file maybe is using!so do not delete!
+            if (fileCount > MIN_NOT_ALLOWED_DELETE_COUNT) {
+                trim(files);
+            }
+        }
     }
 
     protected abstract boolean accept(File file, long totalSize, int totalCount);
@@ -35,15 +44,19 @@ public abstract class LruDiskUsage implements DiskUsage {
     private void trim(List<File> files) {
         long totalSize = countTotalSize(files);
         int totalCount = files.size();
+        DLogger.d("file total count is " + totalCount + ",total size is " + totalSize);
         for (File file : files) {
             boolean accepted = accept(file, totalSize, totalCount);
             if (!accepted) {
+                if (totalCount <= MIN_NOT_ALLOWED_DELETE_COUNT) {
+                    return;
+                }
                 long fileSize = file.length();
                 boolean deleted = file.delete();
                 if (deleted) {
                     totalCount--;
                     totalSize -= fileSize;
-                    DLogger.i("Cache file " + file + " is deleted because it exceeds cache limit");
+                    DLogger.i("Cache file " + file + " is deleted because it exceeds cache limit,remain count is " + totalCount);
                 } else {
                     DLogger.e("Error deleting file " + file + " for trimming cache");
                 }
